frontend/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
bson==0.5.10

